pyproject.toml
[tool.poetry]
name = "bmo-cafe"
version = "0.1.0"
description = ""
authors = ["justanexperiment <justanexperimentpark@gmail.com>"]
readme = "README.md"
packages = [{include = "bmo"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
openai = "^1.25.1"
pyyaml = "^6.0.1"
scipy = "^1.13.0"
tenacity = "^8.2.3"
tiktoken = "^0.6.0"
termcolor = "^2.4.0"
selenium = "^4.20.0"
pyppeteer = "^2.0.0"
pyppeteer-stealth = "^2.7.4"
datetime = "^5.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
